Move data manipulation functions out of index.js and into parseresults function.
